+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   - ลบ Route และ Model ระบบ IsOnline v.2 ออก เพื่อเพิ่มความปลอดภัย
   - เพิ่มการตรวจสอบการเข้าใช้งานจาก PHER+ ก่อนออก token ใช้งาน
   - เปลี่ยน raw sql ของบาง provider เพื่อป้องกัน sql injection
+  - ลบ route และ model ที่ไม่ได้ใช้งาน
+  - Update package
 
 # 3.10.3
 ## Feature